Use Lombok for boilerplate code generation
This commit saves over 100 lines of boilerplate code using Lombok.  I
am quite impressed with Lombok's simple and well-designed set of
features, even if its online documentation is rather lacking at
present.

There are a sea of different Gradle plugins for Lombok.  The one used
in this commit is the one officially endorsed on the Lombok website:
https://projectlombok.org/setup/gradle