Fix ambiguous link targets in the index
Previously, if a hard link referenced a path with multiple revisions,
it was unclear which revision was actually being referenced.  Now,
backups have an associated number that allows precise link target
resolution.

This change is a prerequisite for full garbage collection, as it
enables us to detect when a file may or may not be a link target.